% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFunction.R
\name{donor_Pair_Genotype_Filter}
\alias{donor_Pair_Genotype_Filter}
\title{Donor Pair Genotype Filter}
\usage{
donor_Pair_Genotype_Filter(
  Donor_Pair_Genotype,
  cancer_time,
  method = "flipping",
  rate_Event_Save = 0.01
)
}
\arguments{
\item{Donor_Pair_Genotype}{The donor pair genotype}

\item{cancer_time}{The cancer time}

\item{method}{The method used to find most compatible genotype. Default: flipping}

\item{rate_Event_Save}{The rate used to filter genotype. Default: 0.01}
}
\value{
A filter donor pair genotype list that contain Genotype, is_Have_mustE, is_Used_Time_Estimate
}
\description{
Filter donor pair genotype that used for time estimate
}
