% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFunction.R
\name{estimate_Epsilon_ForMulti}
\alias{estimate_Epsilon_ForMulti}
\title{Estimate epsilon}
\usage{
estimate_Epsilon_ForMulti(
  pat,
  isF,
  isCE = NULL,
  eps = NULL,
  setD,
  eventD,
  multi_thrds = 1L,
  threshold = 0.05,
  threshold1 = 7,
  n_p = 0L,
  T = 10,
  N_iter = 150L,
  lambdaS = 1,
  thrds = 1L,
  R = 1L
)
}
\arguments{
\item{pat}{A matrix for observed genotypes. The col represents the event
index and the row represents a observation whose entries indicate whether
an event has been observed "1" or not "0". Data for different datasets
and event sets is represented using a block matrix}

\item{isF}{A matrix indicates whether the data in dataset[i] and eventset[j]
has been filled "True" or not "False"}

\item{isCE}{A matrix indicates whether calculate epslion of dataset[i] and
eventset[j],"True" or not "False"}

\item{eps}{The initial matrix of epslion}

\item{setD}{The description of dataset index}

\item{eventD}{The description of eventset index}

\item{multi_thrds}{The number of threads for epsilon estimation. Default: 1}

\item{threshold}{The threshold of epslion estimate. Default: 0.05}

\item{threshold1}{The threshold of epslion estimate method selection. Default: 7}

\item{T}{The temperature of simulated annealing algorithm. Default: 10.0}

\item{N_iter}{The max iter number of simulated annealing algorithm. Default: 150}

\item{lambdaS}{Rate of the sampling process. Default: 1.0}

\item{thrds}{The number of threads for parallel execution. Default: 1}

\item{R}{The loop number of guess Function}
}
\value{
A matrix of estimated epsilon
}
\description{
Estimate epsilon for multi set and event
}
