% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/este.R
\name{estimate_Lambda}
\alias{estimate_Lambda}
\title{Estimate Lambda for multiple data and event sets}
\usage{
estimate_Lambda(
  pat,
  poset,
  isF,
  eps,
  setD,
  eventD,
  lambdaS = 1,
  L = 100L,
  sampling = c("forward", "add-remove", "backward", "bernoulli", "pool"),
  maxIter = 100L,
  updateStepSize = 20L,
  tol = 0.001,
  maxLambda = 1e+06,
  neighborhoodDist = 1L,
  thrds = 1L,
  is_update_eps = FALSE,
  seed = NULL
)
}
\arguments{
\item{pat}{A matrix for observed genotypes. The col represents the event
index and the row represents a observation whose entries indicate whether
an event has been observed "1" or not "0". Data for different datasets
and event sets is represented using a block matrix. pat must fill fully.}

\item{poset}{An adjacency matrix represents partial order sets}

\item{isF}{A matrix indicates whether the data in dataset[i] and eventset[j]
has been filled "True" or not "False"}

\item{eps}{The matrix of epslion}

\item{setD}{The description of dataset index}

\item{eventD}{The description of eventset index}

\item{lambdaS}{Rate of the sampling process. Default: 1.0}

\item{L}{Number of samples to be drawn from the proposal in the E-step.
Default: 100}

\item{sampling}{sampling scheme to generate hidden genotypes, OPTIONS:
"forward" - generate occurrence times according to current rate
            parameters, and, from them, generate the hidden genotypes
"add-remove" - generate genotypes, from observed genotypes, using a two-steps
               proposal. First, pick a move uniformly at random: either to
               add or to remove an event. Events are chosen to be removed
               with probability proportional to their rates, and to be added
               with an inverse probability. Second, make genotypes compatible
               with the poset by either adding or removing all events
               incompatible with the poset
"backward" - enumerate all genotypes with Hamming distance
"bernoulli" - generate genotypes from a Bernoulli distribution with success
              probability p = epsilon
"pool" - generate a pool of compatible genotypes according to current rate
         parameters and sample K observations proportional to their Hamming
         distance}

\item{maxIter}{the maximum number of EM iterations. Default: 100}

\item{updateStepSize}{number of EM steps after which the number of samples,
L is doubled. L is increased, if the difference in the parameter estimates
between such consecutive batches is greater than the tolerance level, tol.
Default: 20}

\item{tol}{convergence tolerance for the error rate and the rate parameters.
The EM runs until the difference between the average estimates in the last
two batches is smaller than tol, or until max.iter is reached. Default: 0.001}

\item{maxLambda}{an optional upper bound on the value of the rate
parameters. Default: 1e6}

\item{thrds}{number of threads for parallel execution. Default: 1}

\item{is_update_eps}{whether update eps in lambda estimate program, experience
shows that it is appropriate to make is_update_eps = FALSE. Default: FALSE}

\item{seed}{seed for reproducibility}

\item{neighborhood.dist}{an integer value indicating the Hamming distance
between the observation and the samples generated by "backward" sampling.
This option is used if sampling is set to "backward". Default: 1}
}
\value{
A list of epsilon1, epsilon2, lambda
}
\description{
Lambda estimation for the hidden conjunctive Bayesian
network model (H-CBN), the estimation based on mccbn
(https://github.com/cbg-ethz/MC-CBN)
}
