% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFunction.R
\name{find_most_Compatible_Genotype_by_Flipping}
\alias{find_most_Compatible_Genotype_by_Flipping}
\title{Find most compatible genotype by flipping}
\usage{
find_most_Compatible_Genotype_by_Flipping(
  genotype,
  eps,
  poset,
  max_iter = 10000,
  cores = 1
)
}
\arguments{
\item{genotype}{The true genotype that from the same dataset}

\item{eps}{The epsilon of each event}

\item{poset}{The poset matrix in CBN}

\item{max_iter}{The max loop iter for each sample}

\item{cores}{The parallel cores}
}
\value{
The matrix of donor pair compatible genotype
}
\description{
Find most compatible genotype by flipping compatible event, it is faster than Sampling
}
