% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFunction.R
\name{find_most_Compatible_Genotype_by_Sampling}
\alias{find_most_Compatible_Genotype_by_Sampling}
\title{Find most compatible genotype by sampling}
\usage{
find_most_Compatible_Genotype_by_Sampling(
  sampling_loop,
  sampling_Num_Total,
  max_One_Num,
  poset,
  lambda_s,
  lambdas,
  drivers,
  time_Donor_Geno,
  eps,
  max_Select_Num,
  cache_dir,
  cores1 = 1,
  cores2 = 1
)
}
\arguments{
\item{sampling_loop}{The number of outer sampling loop}

\item{sampling_Num_Total}{The number of sampling compatible genotype in each outer sampling loop,
the total number = sampling_loop * sampling_Num_Total}

\item{max_One_Num}{The max sampling number in each innerloop}

\item{poset}{The poset matrix in CBN}

\item{lambda_s}{The sampling lambda in CBN}

\item{lambdas}{The lambda of each event in CBN}

\item{drivers}{The drivers}

\item{time_Donor_Geno}{The true genotype}

\item{eps}{The epsilon in CBN}

\item{max_Select_Num}{The max number of compatible genotype of each outer sampling loop}

\item{cache_dir}{The cache dir used for save temp files}

\item{cores1}{The parallel cores in Genotype_Generate}

\item{cores2}{The parallel cores in Donor_Pair_Geno_Gerate}
}
\value{
The list of donor pair compatible genotype
}
\description{
Find most compatible genotype by sampling, first create compatible genotype by Genotype_Generate,
then find top k compatible genotype by Donor_Pair_Geno_Gerate
}
