% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/este.R
\name{find_Poset}
\alias{find_Poset}
\title{Find poset set fastly for multiple data and event sets}
\usage{
find_Poset(
  pat,
  isF,
  isCE = NULL,
  eps,
  setD,
  eventD,
  Fine_Tune_Num = 2L,
  thrds = 1L,
  threshold = 1e-04,
  threshold2 = 0.01,
  gmm_min_cluster = 2,
  gmm_convenience_threshold = 0.001,
  gmm_filter_threshold = 0.002,
  n_p = 6L,
  is_update_eps = FALSE,
  T = 10,
  N_iter = 150L,
  lambdaS = 1,
  R = 1L
)
}
\arguments{
\item{pat}{A matrix for observed genotypes. The col represents the event
index and the row represents a observation whose entries indicate whether
an event has been observed "1" or not "0". Data for different datasets
and event sets is represented using a block matrix}

\item{isF}{A matrix indicates whether the data in dataset[i] and eventset[j]
has been filled "True" or not "False"}

\item{isCE}{A matrix indicates whether calculate epslion of dataset[i] and
eventset[j],"True" or not "False"}

\item{eps}{The matrix of epslion}

\item{setD}{The description of dataset index}

\item{eventD}{The description of eventset index}

\item{Fine_Tune_Num}{Fine tune number for more accurate poset, experience shows that
the best number is 2. Default: 2}

\item{thrds}{The number of threads for parallel execution. Default: 1}

\item{threshold}{The threshold for determining partial order relationships, it can filter
out more false positives 1 with the larger value. experience shows that it is appropriate to
threshold <= 0.01, if you set threshold <= 0, the program will search the threshold with gmm,
we recommend setting it to a small value(such as 0.0001) when Fine_Tune_Num >= 2. Default: 0.0001}

\item{threshold2}{The threshold for determining partial order relationships considering the tolerance,
it is appropriate to make threshold2 >= threshold. Default: 0.01}

\item{gmm_min_cluster}{The min cluster number of gmm. Default: 2}

\item{gmm_convenience_threshold}{The value used for stop increasing cluster number, it can
increase the cluster number with the smaller value, experience shows that it is appropriate to
make gmm_convenience_threshold <= 0.005. Default: 0.001}

\item{gmm_filter_threshold}{The value used for searching threshold after gmm, it can increase
the threshold with the larger value, experience shows that it is appropriate to
make gmm_convenience_threshold <= gmm_filter_threshold <= 0.0075. Default: 0.002}

\item{n_p}{The vector contain the number of events during fine-tune. Default: [6, 6]}

\item{is_update_eps}{whether update eps in fine tune, experience
shows that it is appropriate to make is_update_eps = FALSE. Default: FALSE}

\item{T}{The temperature of simulated annealing algorithm. Default: 10.0}

\item{N_iter}{The max iter number of simulated annealing algorithm. Default: 150}

\item{lambdaS}{Rate of the sampling process. Default: 1.0}

\item{R}{The loop number of guess Function}
}
\value{
A matrix of poset
}
\description{
Fast poset set estimation for the hidden conjunctive Bayesian
network model (H-CBN), First determine the partial order relationship
between any two events, and then perform fine-tune
}
