% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFunction.R
\name{fitness_Inference_Patient_Gene}
\alias{fitness_Inference_Patient_Gene}
\title{Fitness Inference Patient Gene Stage2}
\usage{
fitness_Inference_Patient_Gene(
  prior_data,
  sigma_p_square = 1e-04,
  max_iter = 10000,
  cellular_clock = "censored_Rate",
  target_recept_rate = 0.234,
  f_b_scaling_update_strength = 10,
  r_update_sigma = 0.01,
  show_size = 1000,
  seed = NULL
)
}
\arguments{
\item{prior_data}{The prior data from stage1}

\item{sigma_p_square}{The initial variance of the proposed distribution in the adaptive MCMC, default 0.0001}

\item{max_iter}{The max sampling number, default 10000}

\item{cellular_clock}{The colname that used as cellular_clock in sample_Patient, default "censored_Rate"}

\item{target_recept_rate}{The target recept rate in the adaptive MCMC, default 0.234}

\item{f_b_scaling_update_strength}{The scaling update strength in the adaptive MCMC, default 10}

\item{r_update_sigma}{The sigma of the proposed distribution in the gene update stage, default 0.01}

\item{show_size}{The show size, default 1000}

\item{seed}{The seed}
}
\value{
The list of fitness_Inference_Input_Gene, sample_Patient, gene_Temp, driver_gene_Temp,
         fp_bp_Sampling, accepted_fp_bp_Sampling, fp_bp_scaling_Sampling, data_List, gene_List,
         pi_alpha, pi_beta, pi_Sampling, scaleNum, driver_gene_Temp_t_min
}
\description{
Inference fitness for patients with gene neoantigen data and fixed prior
}
