% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFunction.R
\name{fitness_Inference_Patient_Gene_Prior}
\alias{fitness_Inference_Patient_Gene_Prior}
\title{Fitness Prior Inference Patient Gene Stage1}
\usage{
fitness_Inference_Patient_Gene_Prior(
  fitness_Inference_Input_Gene,
  sample_Patient,
  driver_genes,
  min_Patient = 4,
  sigma_p_square = 1e-04,
  neighbor_size = 10,
  history_num = 100,
  history_skip = 10,
  history_size = 25000,
  knn_used_in_history_size = 20000,
  gene_history_num = 100,
  gene_history_size = 25000,
  gene_used_in_history_size = 20000,
  max_iter = 40000,
  cellular_clock = "censored_Rate",
  target_recept_rate = 0.234,
  f_b_scaling_update_strength = 10,
  r_update_sigma = 0.01,
  show_size = 1000,
  seed = NULL
)
}
\arguments{
\item{fitness_Inference_Input_Gene}{The fitness inference gene input data}

\item{sample_Patient}{The patient info}

\item{driver_genes}{The driver genes in CBN that don't need to update the parameters}

\item{min_Patient}{The genes that will be used at least min_Patient, default 4}

\item{sigma_p_square}{The initial variance of the proposed distribution in the adaptive MCMC, default 0.0001}

\item{neighbor_size}{The neighbor size in knn prior stage, default 10}

\item{history_num}{The size of history data in history prior stage, default 100}

\item{history_skip}{The thining num in history prior stage, default 10}

\item{history_size}{The total size before history prior stage, default 25000}

\item{knn_used_in_history_size}{The size of knn prior stage used for history prior stage, default 20000}

\item{gene_history_num}{The gene size of history data in history prior stage, default 100}

\item{gene_history_size}{The total gene size before history prior stage, default 25000}

\item{gene_used_in_history_size}{The gene size of knn prior stage used for history prior stage, default 20000}

\item{max_iter}{The max sampling number, default 40000}

\item{cellular_clock}{The colname that used as cellular_clock in sample_Patient, default "censored_Rate"}

\item{target_recept_rate}{The target recept rate in the adaptive MCMC, default 0.234}

\item{f_b_scaling_update_strength}{The scaling update strength in the adaptive MCMC, default 10}

\item{r_update_sigma}{The sigma of the proposed distribution in the gene update stage, default 0.01}

\item{show_size}{The show size, default 1000}

\item{seed}{The seed}
}
\value{
The list of fitness_Inference_Input_Gene, sample_Patient, gene_Temp, driver_gene_Temp,
         fp_bp_Sampling, accepted_fp_bp_Sampling, fp_bp_scaling_Sampling, data_List, gene_List,
         pi_alpha, pi_beta, pi_Sampling, scaleNum, driver_gene_Temp_t_min
}
\description{
Inference fitness for patients with gene neoantigen data
}
