% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFunction.R
\name{Gamma_Parameter_GS}
\alias{Gamma_Parameter_GS}
\title{Gamma Parameter GS}
\usage{
Gamma_Parameter_GS(
  curr.alpha,
  curr.beta,
  log.hyper.p,
  hyper.q,
  hyper.r,
  hyper.s,
  x,
  gamma.sf
)
}
\arguments{
\item{curr.alpha}{The current alpha}

\item{curr.beta}{The current beta}

\item{log.hyper.p}{The log hyperparameter of p}

\item{hyper.q}{The hyperparameter of q}

\item{hyper.r}{The hyperparameter of r}

\item{hyper.s}{The hyperparameter of s}

\item{x}{The current sampling data}

\item{gamma.sf}{The scaling perparameter for exploration}
}
\value{
A array of the new.alpha, new.beta, accepted
}
\description{
Gamma Parameter Gibbs Sampling
}
