% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFunction.R
\name{GammaCluster}
\alias{GammaCluster}
\title{Gamma Cluster}
\usage{
GammaCluster(data, rate_cluster, max_iter = 10000, tol = 1e-08)
}
\arguments{
\item{data}{The vector of input data}

\item{rate_cluster}{The rate of cluster}

\item{max_iter}{The max update iters}

\item{tol}{The tolerance}
}
\value{
A list that contain the estimation results of weights, shape params, rate params, responsibilities, cluster
}
\description{
Gamma Cluster of 1 dim data based on GMM
}
