% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFunction.R
\name{genotype_SolveQP}
\alias{genotype_SolveQP}
\title{genotype SolveQP}
\usage{
genotype_SolveQP(
  posetCluster,
  age,
  donor,
  cancer_time,
  Geno,
  driver_genes_len,
  filter_num,
  rateT,
  posetClusterWeight,
  isWeight = TRUE
)
}
\arguments{
\item{posetCluster}{The list of different subpathway}

\item{age}{The age of the donor}

\item{donor}{The id of the donor}

\item{cancer_time}{The cancer time}

\item{Geno}{The compatible genotype}

\item{driver_genes_len}{The length(driver_genes)}

\item{filter_num}{The number matrix of chr time event}

\item{rateT}{The list of low and up bound rate of each event}

\item{isWeight}{If isWeight = true, use filter num to weighted average}
}
\value{
A array of the best z1 and z2
}
\description{
find the best z1 and z2 by SolveQP based on genotype
}
