% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFunction.R
\name{get_Res_From_Sampling}
\alias{get_Res_From_Sampling}
\title{Get Results From Sampling}
\usage{
get_Res_From_Sampling(MH_Res, start, end, skip_size)
}
\arguments{
\item{MH_Res}{The result of MH-Sampling}

\item{start}{The start of useful data}

\item{end}{The end of useful data}

\item{skip_size}{The skip size}
}
\value{
A list of alpha, beta, lambda, donor, censored_Time, time_Sampling
}
\description{
get res from sampling
}
