% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFunction.R
\name{indexToRange}
\alias{indexToRange}
\title{Index To Target}
\usage{
indexToRange(arr, index)
}
\arguments{
\item{arr}{A array}

\item{index}{A number}
}
\value{
A range array
}
\description{
Find the range condition on arr and index
}
