% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/este.R
\name{is_Compatible}
\alias{is_Compatible}
\title{Large scale continuous time conjunctive Bayesian network (lsctCBN)
for multiple data and event sets}
\usage{
is_Compatible(genotype, poset)
}
\arguments{
\item{genotype}{A matrix}

\item{poset}{A matrix}
}
\description{
Determine whether the genotype is compatible with the poset
}
