% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFunction.R
\name{Likelihood_Beta}
\alias{Likelihood_Beta}
\title{Likelihood Beta}
\usage{
Likelihood_Beta(x, param)
}
\arguments{
\item{x}{The data}

\item{param}{The parameters}
}
\value{
A number of likelihood
}
\description{
calculate the likelihood of Beta distribution
}
