% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFunction.R
\name{MetropolisHastings_Beta}
\alias{MetropolisHastings_Beta}
\title{MetropolisHastings Beta}
\usage{
MetropolisHastings_Beta(priorParameters, x, update_sigma, start_pos)
}
\arguments{
\item{priorParameters}{The prior parameters}

\item{x}{The data}

\item{update_sigma}{The update sigma}

\item{start_pos}{The start position}
}
\value{
The param
}
\description{
Metropolis-Hastings sampling of Beta distribution
}
