% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFunction.R
\name{MH_Pretreatment_based_on_flipping}
\alias{MH_Pretreatment_based_on_flipping}
\title{MH Pretreatment based on flipping}
\usage{
MH_Pretreatment_based_on_flipping(
  filter_Donor_Pair_Genotype,
  time_temp,
  cancer_time,
  filter_num,
  rateT,
  driver_genes,
  isWeight,
  cBioCli,
  topo_path,
  parent_Set,
  gammaClusterRes
)
}
\arguments{
\item{filter_Donor_Pair_Genotype}{The filter donor pair genotype}

\item{time_temp}{The time temp}

\item{cancer_time}{The cancer time}

\item{filter_num}{The number matrix of chr time event}

\item{rateT}{The list of low and up bound rate of each event}

\item{driver_genes}{The driver genes}

\item{isWeight}{If isWeight = true, use filter num to weighted average}

\item{cBioCli}{The cBioCli}

\item{topo_path}{The topo path}

\item{parent_Set}{The parent set}
}
\value{
A array of the best z1 and z2
}
\description{
MH Pretreatment based on flipping
}
