% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFunction.R
\name{MH_Sampling_based_on_flipping}
\alias{MH_Sampling_based_on_flipping}
\title{MH Sampling based on flipping}
\usage{
MH_Sampling_based_on_flipping(
  Donor_Pair_Geno_Last,
  saveFile,
  isWeight = TRUE,
  iterations = 10000,
  g.sf = 10,
  maxSampleIter = 100,
  errorTolerance = 0.01,
  errorToleranceGrowthRate = 1.1,
  cores1 = 6,
  cores2 = 10,
  show_size = 1000,
  seed = 1234
)
}
\arguments{
\item{Donor_Pair_Geno_Last}{The result from MH_Pretreatment_based_on_flipping}

\item{saveFile}{The file used to save result}

\item{isWeight}{If isWeight = true, use filter num to weighted average}

\item{iterations}{The sampling iterations}

\item{g.sf}{The scaling perparameter for exploration}

\item{maxSampleIter}{The max sample iterations of reject sampling}

\item{errorTolerance}{The error tolerance of reject sampling}

\item{errorToleranceGrowthRate}{The error tolerance growth rate of reject sampling}

\item{cores1}{The parallel cores of sampling for chr time contain data}

\item{cores2}{The parallel cores of sampling for no chr time contain data}

\item{show_size}{The show size}

\item{seed}{The seed}
}
\value{
A list of alpha, beta, lambda, accepted, chr_Time_Contain, no_chr_Time_Contain, time_Sampling
}
\description{
MH Sampling based on flipping
}
