% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFunction.R
\name{MhParameterProposal_Beta}
\alias{MhParameterProposal_Beta}
\title{Mh Parameter Proposal Beta}
\usage{
MhParameterProposal_Beta(update_sigma, old_params)
}
\arguments{
\item{update_sigma}{The update sigma}

\item{old_params}{The old params}
}
\value{
The new params
}
\description{
MH parameter proposal of Beta distribution
}
