% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFunction.R
\name{parent_Set_Help1}
\alias{parent_Set_Help1}
\title{parent Set Help1}
\usage{
parent_Set_Help1(poset)
}
\arguments{
\item{poset}{The poset matrix in CBN}
}
\value{
A list contain the parent set of each event
}
\description{
find parent set for each event
}
