% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFunction.R
\name{PosteriorDraw_Mvnormal_for_1}
\alias{PosteriorDraw_Mvnormal_for_1}
\title{Posterior Draw Mvnormal for one}
\usage{
PosteriorDraw_Mvnormal_for_1(priorParameters, x)
}
\arguments{
\item{priorParameters}{The prior parameters}

\item{x}{The data}
}
\value{
A list of mu, sig
}
\description{
sample one posterior from mvnormal distribution with simple solution for numerical error
}
