% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFunction.R
\name{PosteriorParameters_Mvnormal}
\alias{PosteriorParameters_Mvnormal}
\title{Posterior Parameters Mvnormal}
\usage{
PosteriorParameters_Mvnormal(priorParameters, x)
}
\arguments{
\item{priorParameters}{The prior parameters}

\item{x}{The data}
}
\value{
A list of mu_n, t_n, kappa_n, nu_n
}
\description{
sample posterior parameters from mvnormal distribution
}
