% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFunction.R
\name{PriorDensity_Beta}
\alias{PriorDensity_Beta}
\title{Prior Density Beta}
\usage{
PriorDensity_Beta(priorParameters, param)
}
\arguments{
\item{priorParameters}{The priorParameters}

\item{param}{The param}
}
\value{
A number of prior density
}
\description{
calculate the prior density of Beta distribution
}
