% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFunction.R
\name{random_Poset}
\alias{random_Poset}
\title{Random Poset}
\usage{
random_Poset(p, graph_density = 0.15, trans_reduced = TRUE)
}
\arguments{
\item{p}{The num of event}

\item{graph_density}{The density of graph}

\item{trans_reduced}{Add the transitive reduced poset to the list}
}
\value{
A poset
}
\description{
Generate a random poset
}
