% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFunction.R
\name{rateTimeHelp}
\alias{rateTimeHelp}
\title{Rate Time Help}
\usage{
rateTimeHelp(lambda_s, lambdas, poset)
}
\arguments{
\item{lambda_s}{The sampling lambda in CBN}

\item{lambdas}{The lambda of each event in CBN}

\item{poset}{The poset matrix in CBN}
}
\value{
A list that contain the time's low and up bound of each event
}
\description{
Convert the lambda to the time bound based on lambdaS
}
