% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/este.R
\name{sample_Age_T}
\alias{sample_Age_T}
\title{Large scale continuous time conjunctive Bayesian network (lsctCBN)
for multiple data and event sets}
\usage{
sample_Age_T(data, alpha, beta, lambda, maxSampleIter, rateE1, thrds, seed)
}
\arguments{
\item{data}{A data matrix}

\item{alpha}{Alpha parameter of exp distribution}

\item{beta}{Beta parameter of exp distribution}

\item{lambda}{Parameter of exp distribution}

\item{maxSampleIter}{Max ssample iter}

\item{rateE1}{Tolerated amplification rate}

\item{thrds}{Parallel thrds}
}
\description{
Sample time for censor age
}
