% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/este.R
\name{sample_Age_TLW}
\alias{sample_Age_TLW}
\title{Large scale continuous time conjunctive Bayesian network (lsctCBN)
for multiple data and event sets}
\usage{
sample_Age_TLW(data, alpha, beta, lambda, thrds)
}
\arguments{
\item{data}{A data matrix}

\item{alpha}{Alpha parameter of exp distribution}

\item{beta}{Beta parameter of exp distribution}

\item{lambda}{Parameter of exp distribution}

\item{thrds}{Parallel thrds}

\item{maxSampleIter}{Max ssample iter}

\item{rateE1}{Tolerated amplification rate}
}
\description{
Sample time for censor age (lightweight)
}
