% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselineTest.R
\name{simulation_Data_Generate}
\alias{simulation_Data_Generate}
\title{Generate simulation data}
\usage{
simulation_Data_Generate(
  numEventArray,
  numSampleArray,
  lambdaSampling = 1,
  lambdaSamplingScaling = 3,
  baseEpsilon = 0.05,
  epsilonSamplingRate = 0.5,
  graphDensity = 0.2
)
}
\arguments{
\item{numEventArray}{A array describe the event number in each event type}

\item{numSampleArray}{A array describe the set number in each set type}

\item{lambdaSampling}{The value of lambdas. Default: 1}

\item{lambdaSamplingScaling}{The value of scaling which used to sample lambdas. Default: 3}

\item{baseEpsilon}{The value of baseEpsilon. Default: 0.05}

\item{epsilonSamplingRate}{The value of rate which used to sample epsilon. Default: 0.5}

\item{graphDensity}{The value which used to control the sparsity of poset. Default: 0.2}
}
\description{
Generate simulation data
}
