% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFunction.R
\name{targetToIndex}
\alias{targetToIndex}
\title{Target to index}
\usage{
targetToIndex(arr, target)
}
\arguments{
\item{arr}{A array}

\item{target}{A number}
}
\value{
A number
}
\description{
Find the max index which sum(arr|index|) >= target
}
