% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseFunction.R
\name{topological_Sort}
\alias{topological_Sort}
\title{Topological Sort}
\usage{
topological_Sort(poset)
}
\arguments{
\item{poset}{A matrix for poset}
}
\value{
A sorted list
}
\description{
Perform topological sort with poset
}
